requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.1.1
reportlab==4.0.4
icalendar==5.0.7
openpyxl==3.1.2
PyMySQL==1.1.0
cryptography==41.0.4